Create interface for applications operations, move and rename ApplicationsOperations and move to other package
